timorblog/requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
django-filter==1.0.4
django-uuslug==1.1.8
djangorestframework==3.6.3
olefile==0.44
Pillow==4.2.1
python-slugify==1.2.4
pytz==2017.2
Unidecode==0.4.21